+++ src/ZeppelinForms/Forms/Controls/CheckBox.cs
@@ -52,15 +52,13 @@ public class CheckBox : InteractiveControl, ITextElement
     public event EventHandler? CheckedChanged;
 
     public string? Text { get; set; }
-    public HorizontalContentAlignment ContentAlign { get; set; } = HorizontalContentAlignment.Left;
-    public VerticalContentAlignment ContentVerticalAlign { get; set; } = VerticalContentAlignment.Center;
     public Color TextColor { get; set; } = Colors.Black;
 
     public Color BoxBorderColor { get; set; } = Colors.Black;
     public Color BoxBackground { get; set; } = Colors.White;
     public Color CheckColor { get; set; } = new Color(255, 0x0D, 0x6E, 0xFD);
-    public HorizontalContentAlignment HorizontalContentAlign { get; set; }
-    public VerticalContentAlignment VerticalContentAlign { get; set; }
+    public HorizontalContentAlignment HorizontalContentAlign { get; set; } = HorizontalContentAlignment.Left;
+    public VerticalContentAlignment VerticalContentAlign { get; set; } = VerticalContentAlignment.Center;
 
     public CheckBox()
     {
@@ -121,7 +119,7 @@ public class CheckBox : InteractiveControl, ITextElement
             new Point(content.X + BoxSize + Gap, content.Y),
             new Size(Math.Max(0, content.Width - BoxSize - Gap), content.Height));
 
-        g.DrawText(Text, textRect, TextColor, EffectiveFont, ContentAlign, ContentVerticalAlign);
+        g.DrawText(Text, textRect, TextColor, EffectiveFont, this.HorizontalContentAlign, this.VerticalContentAlign);
     }
 
     private static void DrawCheckMark(Graphics g, Rectangle box)